app/requirements.txt
flask==3.0.3
prometheus-flask-exporter==0.22.4
mysql-connector-python